Delete invalid filename for Netlify deploy
Removed the file named '# ![NGO Logo](logo.png)' from the repository because it contained special characters (#, !, [, ], ( )) that are not allowed in deployed filenames. This was causing Netlify deploy failures. The valid logo file 'logo.png' remains in the repository for use in the project. This cleanup ensures successful deployment of the site on Netlify.